+++ Sandboxie/apps/start/start.cpp
@@ -92,7 +92,6 @@ BOOL auto_select_default_box = FALSE;
 WCHAR *StartMenuSectionName = NULL;
 BOOL run_silent = FALSE;
 BOOL keep_alive = FALSE;
-BOOL dont_start_sbie_ctrl = FALSE;
 BOOL hide_window = FALSE;
 BOOL wait_for_process = FALSE;
 BOOLEAN layout_rtl = FALSE;
@@ -577,16 +576,6 @@ BOOL Parse_Command_Line(void)
             env_value_x[env_value_len] = L'\0';
 
             SetEnvironmentVariable(env_name_x, env_value_x);
-
-        //
-        // Command line switch /nosbiectrl
-        //
-
-        } else if (_wcsnicmp(cmd, L"nosbiectrl", 10) == 0) {
-
-            cmd += 10;
-
-            dont_start_sbie_ctrl = TRUE;
 
         //
         // Command line switch /reload